Separate `seed-data` module
